Time: 2 ms (82.64%), Space: 44.7 MB (77.34%) - LeetHub

+++ 0383-ransom-note/0383-ransom-note.java
@@ -1,6 +1,6 @@
 class Solution {
     public boolean canConstruct(String ransomNote, String magazine) {
-        int[] chars  = new int[26];
+        short[] chars  = new short[26];
         for( char c : magazine.toCharArray()){
             chars[c-'a']++;
         }
